Random-Password-Generator/requirements.txt
Python 3.10
PyQt5 
random
string
re
requests
mysql.connector
Custom_Widgets (for PyQt5)